style: arrange user helper & topFollowedUser
