+++ controllers/pages/user-controller.js
@@ -1,15 +1,122 @@
 const bcrypt = require('bcryptjs')
 
-const { User, Tweet, Reply, Like, sequelize } = require('../../models')
+const { User, Tweet, Reply, Like, Followship, sequelize } = require('../../models')
+const { Op } = require('sequelize')
 const helpers = require('../../_helpers')
-const recommendUserHelper = require('../../helpers/recommand-followship-helper')
-const userHelper = require('../../helpers/user-helper')
 const userService = require('../../service/user-services')
 const errorHandler = require('../../helpers/errors-helpers')
 
 const INPUT_LENGTH_JS = 'inputLength.js'
 const USER_PAGE_JS = 'userPage.js'
 const CHECK_PASSWORD_JS = 'checkPassword.js'
+
+const userHelper = {
+  getUserInfo: async req => {
+    return await User.findByPk(req.params.id, {
+      attributes: {
+        include: [
+          [sequelize.literal('( SELECT COUNT(*) FROM Followships WHERE Followships.follower_id = User.id)'), 'followingCount'],
+          [sequelize.literal('( SELECT COUNT(*) FROM Followships WHERE Followships.following_id = User.id)'), 'followerCount'],
+          [sequelize.literal('( SELECT COUNT(*) FROM Tweets WHERE Tweets.user_id = User.id)'), 'tweetsCount'],
+          [sequelize.literal(
+              `(SELECT COUNT(*) FROM Followships
+                WHERE Followships.follower_id = ${helpers.getUser(req).id}
+                AND Followships.following_id = ${req.params.id}
+              )`), 'isFollowed']
+        ]
+      },
+      raw: true,
+      nest: true
+    })
+  },
+
+  getFollowings: async req => {
+    const userId = req.params.id
+    const userWithfollowings = await User.findByPk(userId, {
+      include: [
+        {
+          model: User,
+          as: 'Followings',
+          attributes: [
+            'id',
+            'name',
+            'account',
+            'avatar',
+            'introduction',
+            [sequelize.literal(
+              `(SELECT COUNT(*) FROM Followships
+                WHERE Followships.follower_id = ${userId}
+                AND Followships.following_id = Followings.id
+              )`), 'isFollowed'] // 查看此User是否已追蹤
+          ],
+          through: {
+            model: Followship,
+            attributes: ['createdAt']
+          }
+        }
+      ],
+      order: [[{ model: User, as: 'Followings' }, { model: Followship }, 'createdAt', 'DESC']]
+
+    })
+    return userWithfollowings.toJSON()
+  },
+
+  getFollowers: async req => {
+    const userId = req.params.id
+    const userWithfollowers = await User.findByPk(userId, {
+      include: [
+        {
+          model: User,
+          as: 'Followers',
+          attributes: [
+            'id',
+            'name',
+            'account',
+            'avatar',
+            'introduction',
+            [sequelize.literal(
+              `(SELECT COUNT(*) FROM Followships
+                WHERE Followships.follower_id = ${userId}
+                AND Followships.following_id = Followers.id
+              )`), 'isFollowed'] // 查看此User是否已追蹤
+          ],
+          through: {
+            model: Followship,
+            attributes: ['createdAt']
+          }
+        }
+      ],
+      order: [[{ model: User, as: 'Followers' }, { model: Followship }, 'createdAt', 'DESC']]
+
+    })
+    return userWithfollowers.toJSON()
+  },
+
+  topFollowedUser: req => {
+    return User.findAll({
+      where: {
+        id: { [Op.ne]: helpers.getUser(req).id }, // 不要出現登入帳號
+        role: { [Op.ne]: 'admin' } // admin不推薦, ne = not
+      },
+      attributes: {
+        include: [
+          // 使用 sequelize.literal 創建一個 SQL 子查詢來計算帖子數量
+          [sequelize.literal('(SELECT COUNT(*) FROM Followships WHERE Followships.following_id = User.id)'), 'followerCount'], // User不要加s, 坑阿！
+          // req.user是追別人的,  findAll的user是被追的人
+          [sequelize.literal(
+            `(SELECT COUNT(*) FROM Followships
+              WHERE Followships.follower_id = ${helpers.getUser(req).id}
+              AND Followships.following_id = User.id
+            )`), 'isFollowed'] // 查看此User是否已追蹤
+        ]
+      },
+      limit: 10,
+      order: [['followerCount', 'DESC']],
+      raw: true,
+      nest: true
+    })
+  }
+}
 
 const userController = {
   /* user登入 */
@@ -154,7 +261,7 @@ const userController = {
         nest: true
       })
 
-      const recommendUser = await recommendUserHelper.topFollowedUser(req)
+      const recommendUser = await userHelper.topFollowedUser(req)
 
       return res.render('user/tweets', {
         route: 'user',
@@ -197,7 +304,7 @@ const userController = {
 
   getFollowings: async (req, res, next) => {
     const followings = await userHelper.getFollowings(req)
-    const recommendUser = await recommendUserHelper.topFollowedUser(req)
+    const recommendUser = await userHelper.topFollowedUser(req)
 
     return res.render('user/followings', {
       userTab: 'followings',
@@ -208,7 +315,7 @@ const userController = {
 
   getFollowers: async (req, res, next) => {
     const followers = await userHelper.getFollowers(req)
-    const recommendUser = await recommendUserHelper.topFollowedUser(req)
+    const recommendUser = await userHelper.topFollowedUser(req)
 
     return res.render('user/followers', {
       userTab: 'followers',
@@ -226,7 +333,7 @@ const userController = {
       const viewingUser = await userHelper.getUserInfo(req)
       if (!viewingUser) throw new errorHandler.UserError("User didn't exist")
 
-      const recommendUser = await recommendUserHelper.topFollowedUser(req)
+      const recommendUser = await userHelper.topFollowedUser(req)
 
       const tweets = await Tweet.findAll({
         include: [
@@ -273,7 +380,7 @@ const userController = {
       const viewingUser = await userHelper.getUserInfo(req)
       if (!viewingUser) throw new errorHandler.UserError("User didn't exist")
 
-      const recommendUser = await recommendUserHelper.topFollowedUser(req)
+      const recommendUser = await userHelper.topFollowedUser(req)
 
       const tweets = await Tweet.findAll({
         include: [
@@ -312,4 +419,7 @@ const userController = {
   }
 }
 
-module.exports = userController
+module.exports = {
+  userController,
+  userHelper
+}
