+++ routes/index.js
@@ -6,7 +6,7 @@ const tweet = require('./modules/tweet')
 const user = require('./modules/user')
 const api = require('./modules/api')
 
-const userController = require('../controllers/pages/user-controller')
+const { userController } = require('../controllers/pages/user-controller')
 const followController = require('../controllers/pages/follow-controller')
 
 const errorHandler = require('../middlewares/error-handler')
